[ADD] init listener structure by using bind() and listen(), [MOD] check of port, bind is watching for this already. Fill of structure in init_listener.c file
