+++ includes/socket.h
@@ -79,4 +79,16 @@ struct socket_s {
  */
 uint32_t create_socket_file_descriptor(in_port_t port, struct socket_s *socket_controller);
 
+/**
+ * @brief Initialize listener structure by fill the structure,
+ * binding the file descriptor and listening it to manage incoming connexion.
+ * 
+ * @param socket_fd The socket fd created by socket function
+ * @param port The port listener
+ * @param address The address of the data socket connexion
+ * @param socket_controller The structure controller
+ * @return uint32_t FAILURE in case of error / SUCCESS if all has been done successfuly
+ */
+uint32_t initialize_socket_controller(int32_t socket_fd, in_port_t port, struct sockaddr_in *address, struct socket_s *socket_controller);
+
 #endif /* !SOCKET_H_ */
